MIPS: Loongson 3: Introduce disable_unused_cpus() to save power.
When use maxcpus=n to limit the number of cores, the initial offlined
cores are running a infinite loop in firmware and their clocks are not
disabled. For power saving, we take the unused cores up and then down
at boot time. When resume from STR, we do this the same.

Change-Id: I367f08942948c141211eeabb7687a92beb66879d
Signed-off-by: Huacai Chen <chenhc@lemote.com>
Signed-off-by: Hongliang Tao <taohl@lemote.com>
Signed-off-by: Hua Yan <yanh@lemote.com>
Signed-off-by: Zhang Shuangshuang <zhangshuangshuang@ict.ac.cn>