MIPS: Loongson 3: Add CPU Autoplug support.
CPU Autoplug is used for power saving. Loongson 3 supports CPUFreq, but
the clock frequency of each core is always the same (cannot be adjusted
independently). If there is a single-thread CPU-intensive program, we
cannot use frequency scaling for power saving (if we do, performance of
the program will slowdown). Therefore, we introduce CPU Autoplug to
adjust the number of online cores based on the global CPU load.

Change-Id: Id54d2eea052efdad4b55f8f98e847618213b809f
Signed-off-by: Huacai Chen <chenhc@lemote.com>
Signed-off-by: Hongliang Tao <taohl@lemote.com>
Signed-off-by: Hua Yan <yanh@lemote.com>
Signed-off-by: Zhang Shuangshuang <zhangshuangshuang@ict.ac.cn>