MIPS: Loongson: emulate CPUCFG instruction on older cores
CPUCFG is the new instruction for querying processor characteristics in
Loongson land, much like CPUID of x86. Since the instruction is
supposedly designed to provide a unified way to do feature detection on
Loongson (without having to, for example, parse /proc/cpuinfo which is
too heavyweight), it is important to provide compatibility for older
cores without native support. Fortunately, most of the fields can be
synthesized without changes to semantics. Performance is not really big
a concern, because feature detection logic is not expected to be invoked
very often in typical userland applications.

This is lightly based on Loongson's work on their Linux 3.10 fork, for
being the authority on the right feature flags to fill in, where things
aren't otherwise discoverable.

Signed-off-by: WANG Xuerui <git@xen0n.name>
Cc: Huacai Chen <chenhc@lemote.com>
Cc: Jiaxun Yang <jiaxun.yang@flygoat.com>
Cc: Tiezhu Yang <yangtiezhu@loongson.cn>
Cc: linux-mips@vger.kernel.org