swiotlb: Keep offsets when mapping data
NVMe driver and other applications depend on the data offset
to operate correctly. Currently when unaligned data is mapped via
SWIOTLB, the data is mapped as slab aligned with the SWIOTLB.

When booting with swiotlb=force option and using NVMe as interface,
running mkfs.xfs fails because of the unalignment issue.

We need to make sure the mapped data preserves its offset of
the orginal address.

Signed-off-by: Jianxiong Gao <jxgao@google.com>
Acked-by: David Rientjes <rientjes@google.com>
Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>