LoongArch: Clean up the architectural interrupt definitions
While interrupts are assigned ECodes `64 + interrupt number`, all
existing use sites of interrupt numbers want the 64 subtracted.
Re-arrange the definitions so that the actual interrupt number is used
everywhere, and make EXCCODE_INT_END inclusive as it is more intuitive
that way.

While at it, document an ISA manual erratum that is confirmed by Huacai
but not yet officially fixed in the official (Chinese) version of the
manual, namely the total number of architectural interrupts should be
(at least) 14 instead of 13. Relevant bitfields and masks have been
adjusted as well.

Signed-off-by: WANG Xuerui <git@xen0n.name>