arm64: efi: Remap kernel image after moving it in memory
UEFI now has a memory attributes protocol that permits callers to
manipulate permissions of memory mappings directly, and firmware
implementations that expose it may no longer map freshly allocated pages
both writable and executable at the same time, as this is widely
recognized as being a terrible idea for security.

When the image executes from the same place where it was loaded by the
firmware, we can assume that all code and data regions are mapped with
the appropriate permissions. However, if the image is moved around in
memory, either for randomization purposes, or if the original placement
is unsuitable for some reason, we should inform the firmware which part
if executable and which part is writable before handing over to it.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>