ovl: move ovl_file_accessed() to aio completion
Refactor the io completion helpers to handlers with clear hierarchy:
- ovl_aio_rw_complete() is called on completion of submitted aio
 `- ovl_aio_cleanup() is called after any aio submission attempt
  `- ovl_rw_complete() is called after any io attempt

Move ovl_copyattr() and ovl_file_accessed() to the common helper
ovl_rw_complete(), so that they are called after aio completion.

Note that moving ovl_file_accessed() changes touch_atime() therein to
be called with mounter credentials in the sync read case.

It does not seem to matter with which credentials touch_atime() is called.
If it did matter, we would have needed to override to mounter credentials
in ovl_update_time() which calls touch_atime() on upper dentry.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>